[Spanify] `span` size/offset to conform to `size_t` (#26822)
There's some tightening around values passed into span, with
`StrictNumeric<size_t>` being enforced. This is mostly unimportant when
dealing with compile-time values, as they can be validated on the spot,
however it does affect several places where the value is unknown at
runtime, and a signed value was being used.

Chromium changes:
https://chromium.googlesource.com/chromium/src/+/4937ad9ae64d9d3f05d13c9d041769432a12281b

```
commit 4937ad9ae64d9d3f05d13c9d041769432a12281b
Author: Peter Kasting <pkasting@chromium.org>
Date:   Tue Nov 26 22:58:16 2024 +0000

    Use StrictNumeric<size_t> for more span methods.

    Changes subspan(), split_at(), at(), and get_at() to match pre-existing
    usage in constructors, first(), last(), etc.

    Bug: 364987728
```

Resolves https://github.com/brave/brave-browser/issues/42627

+++ components/brave_wallet/common/brave_wallet_constants.h
@@ -8,11 +8,11 @@
 
 namespace brave_wallet {
 
-inline constexpr int kSolanaKeypairSize = 64;
-inline constexpr int kSolanaSignatureSize = 64;
-inline constexpr int kSolanaPrikeySize = 32;
-inline constexpr int kSolanaPubkeySize = 32;
-inline constexpr int kSolanaHashSize = 32;
+inline constexpr size_t kSolanaKeypairSize = 64;
+inline constexpr size_t kSolanaSignatureSize = 64;
+inline constexpr size_t kSolanaPrikeySize = 32;
+inline constexpr size_t kSolanaPubkeySize = 32;
+inline constexpr size_t kSolanaHashSize = 32;
 // 1232 = 1280(IPv6 minimum MTU) - 40(size of the IPv6 header) - 8(size of the
 // fragment header)
 inline constexpr size_t kSolanaMaxTxSize = 1232;
